Moved some importer files around to reuse it in SAM
