+++ src/groovy/dbnp/importer/impl/AssaysImporter.groovy
@@ -1,5 +1,6 @@
-package dbnp.importer
+package dbnp.importer.impl
 
+import dbnp.importer.*
 import dbnp.authentication.SecUser
 import dbnp.studycapturing.*
 import org.dbnp.gdt.*
@@ -8,13 +9,13 @@ import grails.util.Holders
 /**
  * Defines the interface for an exporter
  */
-public class SamplesImporter extends StudyTemplateEntityImporter<Sample> {
+public class AssaysImporter extends StudyTemplateEntityImporter<Assay> {
     
     /**
      * Returns an identifier that describes this importer
      */
     public String getIdentifier() {
-        "Samples"
+        "Assays"
     }
     
     /**
@@ -22,7 +23,7 @@ public class SamplesImporter extends StudyTemplateEntityImporter<Sample> {
      * Can be used to filter the available importers on a certain type
      */
     public boolean supportsType(String type) {
-        type in [ "clinicaldata", "samples" ]
+        type in [ "clinicaldata", "assays" ]
     }
     
     /**
@@ -30,8 +31,8 @@ public class SamplesImporter extends StudyTemplateEntityImporter<Sample> {
      */
     public Map getLinkToResults(def parameters) {
         [
-            url: [ controller: 'study', action: 'samples', id: parameters.study],
-            label: "Samples for study " + getStudy(parameters)?.title
+            url: [ controller: 'study', action: 'assays', id: parameters.study],
+            label: "Assays for study " + getStudy(parameters)?.title
         ]
     }
     
@@ -39,6 +40,6 @@ public class SamplesImporter extends StudyTemplateEntityImporter<Sample> {
      * Returns an entity object for this TemplateEntity (T)
      */
     public Class getEntity() {
-        Sample
+        Assay
     }
 }
