+++ src/groovy/dbnp/importer/impl/SamplesImporter.groovy
@@ -1,20 +1,21 @@
-package dbnp.importer
+package dbnp.importer.impl
 
+import dbnp.importer.*
 import dbnp.authentication.SecUser
 import dbnp.studycapturing.*
 import org.dbnp.gdt.*
 import grails.util.Holders
 
 /**
- * Importer to store events
+ * Defines the interface for an exporter
  */
-public class EventsImporter extends StudyTemplateEntityImporter<Event> {
+public class SamplesImporter extends StudyTemplateEntityImporter<Sample> {
     
     /**
      * Returns an identifier that describes this importer
      */
     public String getIdentifier() {
-        "Events"
+        "Samples"
     }
     
     /**
@@ -22,7 +23,7 @@ public class EventsImporter extends StudyTemplateEntityImporter<Event> {
      * Can be used to filter the available importers on a certain type
      */
     public boolean supportsType(String type) {
-        type in [ "clinicaldata", "events" ]
+        type in [ "clinicaldata", "samples" ]
     }
     
     /**
@@ -30,8 +31,8 @@ public class EventsImporter extends StudyTemplateEntityImporter<Event> {
      */
     public Map getLinkToResults(def parameters) {
         [
-            url: [ controller: 'study', action: 'design', id: parameters.study],
-            label: "Design for study " + getStudy(parameters)?.title
+            url: [ controller: 'study', action: 'samples', id: parameters.study],
+            label: "Samples for study " + getStudy(parameters)?.title
         ]
     }
     
@@ -39,6 +40,6 @@ public class EventsImporter extends StudyTemplateEntityImporter<Event> {
      * Returns an entity object for this TemplateEntity (T)
      */
     public Class getEntity() {
-        Event
+        Sample
     }
 }
