+++ src/groovy/dbnp/importer/impl/SamplingEventsImporter.groovy
@@ -1,20 +1,21 @@
-package dbnp.importer
+package dbnp.importer.impl
 
+import dbnp.importer.*
 import dbnp.authentication.SecUser
 import dbnp.studycapturing.*
 import org.dbnp.gdt.*
 import grails.util.Holders
 
 /**
- * Defines the interface for an exporter
+ * Importer to store sampling events
  */
-public class AssaysImporter extends StudyTemplateEntityImporter<Assay> {
+public class SamplingEventsImporter extends StudyTemplateEntityImporter<SamplingEvent> {
     
     /**
      * Returns an identifier that describes this importer
      */
     public String getIdentifier() {
-        "Assays"
+        "Sampling Events"
     }
     
     /**
@@ -22,7 +23,7 @@ public class AssaysImporter extends StudyTemplateEntityImporter<Assay> {
      * Can be used to filter the available importers on a certain type
      */
     public boolean supportsType(String type) {
-        type in [ "clinicaldata", "assays" ]
+        type in [ "clinicaldata", "samplingevents" ]
     }
     
     /**
@@ -30,8 +31,8 @@ public class AssaysImporter extends StudyTemplateEntityImporter<Assay> {
      */
     public Map getLinkToResults(def parameters) {
         [
-            url: [ controller: 'study', action: 'assays', id: parameters.study],
-            label: "Assays for study " + getStudy(parameters)?.title
+            url: [ controller: 'study', action: 'design', id: parameters.study],
+            label: "Design for study " + getStudy(parameters)?.title
         ]
     }
     
@@ -39,6 +40,6 @@ public class AssaysImporter extends StudyTemplateEntityImporter<Assay> {
      * Returns an entity object for this TemplateEntity (T)
      */
     public Class getEntity() {
-        Assay
+        SamplingEvent
     }
 }
