+++ src/groovy/dbnp/importer/impl/SubjectsImporter.groovy
@@ -1,5 +1,6 @@
-package dbnp.importer
+package dbnp.importer.impl
 
+import dbnp.importer.*
 import dbnp.authentication.SecUser
 import dbnp.studycapturing.*
 import org.dbnp.gdt.*
